Create index.html
A "Simon Says" style game where the player must first press any key to start the game and then one of four tiles will light up with an accompanying sound effect. The player must click on that tile to copy the pattern. And as you go from level to level, the patterns get more and more complicated. If you get it wrong, a game over message is displayed. You can press any key to restart the game.