Create index.js
A JavaScript program using Node.js and the inquirer and qr-image modules to create a QR code generator. It asks the user for a URL for a website they want a QR code for. It also creates a text file and writes the URL to it using the Node native fs (file system) module.